Add transactionHelper to NFT
Deploy transaction parameter calculation was extracted from the deploy
script into a self standing utility. This simplifies the deploy scripts.
