+++ NFT/scripts/deploy.js
@@ -1,18 +1,7 @@
-const { calcEthereumTransactionParams } = require("@acala-network/eth-providers");
-
-const txFeePerGas = '199999946752';
-const storageByteDeposit = '100000000000000';
+const { txParams } = require("../utils/transactionHelper");
 
 async function main() {
-  const blockNumber = await ethers.provider.getBlockNumber();
-
-  const ethParams = calcEthereumTransactionParams({
-    gasLimit: '21000010',
-    validUntil: (blockNumber + 100).toString(),
-    storageLimit: '640010',
-    txFeePerGas,
-    storageByteDeposit
-  });
+  const ethParams = await txParams();
 
   const [deployer, user] = await ethers.getSigners();
